Adopt a Code of Conduct
I think it's a shame that the second most popular github repository does not have a code of conduct and I intend to fix this. We never needed a code of conduct, we don't want to need a code of conduct and that's why we need a code of conduct.

I submit this as a PR instead of commiting directly to master because I want to allow some debate on this issue. I will merge this PR in a week or so, if there is no good case against it.